pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vla-data-curator"
version = "0.1.0"
description = "Data curation and preprocessing framework for Vision-Language-Action (VLA) model training, with focus on reasoning-aware embodied datasets."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "VLA Curator Contributors" }]

dependencies = [
    # Core scientific
    "numpy>=1.24",
    "Pillow>=10.0",
    "matplotlib>=3.7",
    "scipy>=1.11",

    # Data / serialization
    "datasets>=2.18",          # HuggingFace datasets (ECoT loader)
    "huggingface-hub>=0.22",
    "jsonlines>=4.0",
    "pydantic>=2.5",
    "omegaconf>=2.3",

    # Progress / CLI
    "tqdm>=4.66",
    "rich>=13.0",
    "click>=8.1",
    "typer>=0.9",

    # Utilities
    "tenacity>=8.2",           # Retry / backoff for API calls
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
gemini = ["google-generativeai>=0.5"]
openai = ["openai>=1.20"]
qwen-api = ["openai>=1.20"]         # Qwen via OpenAI-compatible API
qwen-local = ["transformers>=4.40", "accelerate>=0.28", "torch>=2.1"]
bridge = ["tensorflow>=2.14", "tensorflow-datasets>=4.9"]
viz = ["imageio>=2.34", "imageio-ffmpeg>=0.4", "seaborn>=0.13"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "ruff>=0.4",
    "mypy>=1.9",
    "ipykernel>=6.29",
    "ipywidgets>=8.1",
]
all = [
    "vla-data-curator[gemini,openai,qwen-api,bridge,viz,dev]",
]

[project.scripts]
vla-visualize   = "vla_curator.scripts.visualize_ecot:main"
vla-generate    = "vla_curator.scripts.generate_traces:main"
vla-curate      = "vla_curator.scripts.curate_interleaved:main"
vla-validate    = "vla_curator.scripts.validate_dataset:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
